requirements_conda.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
attrs=19.1.0=py36_1
backcall=0.1.0=py36_0
blas=1.0=mkl
bleach=3.1.0=py36_0
blosc=1.16.3=he1b5a44_1
bzip2=1.0.8=h516909a_0
ca-certificates=2019.5.15=0
certifi=2019.6.16=py36_1
cffi=1.11.5=py36he75722e_1
cftime=1.0.3.4=py36hd352d35_1001
curl=7.64.1=hf8cf82a_0
cycler=0.10.0=py36_0
cython=0.29.12=py36he1b5a44_0
dbus=1.13.2=h714fa37_1
decorator=4.4.0=py36_1
defusedxml=0.6.0=py_0
entrypoints=0.3=py36_0
expat=2.2.6=he6710b0_0
fftw3f=3.3.4=2
fontconfig=2.13.0=h9420a91_0
freetype=2.9.1=h8a8886c_1
glib=2.56.2=hd408876_0
gmp=6.1.2=h6c8ec71_1
gst-plugins-base=1.14.0=hbbd80ab_1
gstreamer=1.14.0=hb453b48_1
hdf4=4.2.13=0
hdf5=1.10.2=hc401514_3
icu=58.2=h9c2bf20_1
imageio=2.4.1=py36_0
intel-openmp=2019.0=118
ipykernel=5.1.0=py36h39e3cac_0
ipython=7.5.0=py36h39e3cac_0
ipython_genutils=0.2.0=py36_0
ipywidgets=7.4.2=py36_0
jedi=0.13.3=py36_0
jinja2=2.10.1=py36_0
jpeg=9b=h024ee3a_2
jsonschema=3.0.1=py36_0
jupyter=1.0.0=py36_7
jupyter_client=5.2.4=py36_0
jupyter_console=6.0.0=py36_0
jupyter_core=4.4.0=py36_0
kiwisolver=1.0.1=py36hf484d3e_0
krb5=1.16.3=h05b26f9_1001
libcurl=7.64.1=hda55be3_0
libedit=3.1.20170329=h6b74fdf_2
libffi=3.2.1=he1b5a44_1006
libgcc-ng=9.1.0=hdf63c60_0
libgfortran=3.0.0=1
libgfortran-ng=7.3.0=hdf63c60_0
libnetcdf=4.6.1=h5e45101_3
libpng=1.6.36=hbc83047_0
libsodium=1.0.16=h1bed415_0
libssh2=1.8.2=h22169c7_2
libstdcxx-ng=8.2.0=hdf63c60_1
libtiff=4.0.9=he85c1e1_2
libuuid=1.0.3=h1bed415_2
libxcb=1.13=h1bed415_1
libxml2=2.9.8=h26e45fe_1
lz4-c=1.8.1.2=h14c3975_0
lzo=2.10=h14c3975_1000
markupsafe=1.1.1=py36h7b6447c_0
matplotlib=3.0.3=py36h5429711_0
mdtraj=1.9.3=py36h00575c5_0
mistune=0.8.4=py36h7b6447c_0
mkl=2018.0.3=1
mkl_fft=1.0.10=py36_0
mkl_random=1.0.2=py36_0
mock=3.0.5=py36_0
nbconvert=5.4.1=py36_3
nbformat=4.4.0=py36_0
ncurses=6.1=hf484d3e_1002
netcdf4=1.4.1=py36ha292673_200
nglview=0.6.2.3=py36_0
ninja=1.8.2=py36h6bb024c_1
notebook=5.7.8=py36_0
numexpr=2.6.9=py36h637b7d7_1000
numpy=1.15.4=py36h1d66e8a_0
numpy-base=1.15.4=py36h81de0dd_0
olefile=0.46=py36_0
openmm=7.3.1=py36_cuda92_rc_2
openmmtools=0.18.1=pypi_0
openssl=1.1.1c=h7b6447c_1
pandas=0.24.2=py36hb3f55d8_0
pandoc=2.2.3.2=0
pandocfilters=1.4.2=py36_1
parmed=3.2.0=py36_0
parso=0.4.0=py_0
pcre=8.42=h439df22_0
pexpect=4.7.0=py36_0
pickleshare=0.7.5=py36_0
pillow=5.3.0=py36h34e0f95_0
pip=19.1.1=py36_0
prometheus_client=0.6.0=py36_0
prompt_toolkit=2.0.9=py36_0
ptyprocess=0.6.0=py36_0
pycparser=2.19=py36_0
pygments=2.3.1=py36_0
pyparsing=2.2.2=py36_0
pyqt=5.9.2=py36h05f1152_2
pyrsistent=0.14.11=py36h7b6447c_0
pytables=3.4.4=py36h4f72b40_1
python=3.6.7=h357f687_1005
python-dateutil=2.8.0=py_0
pytorch-cpu=1.1.0=py3.6_cpu_0
pytz=2019.1=py_0
pyyaml=5.1.1=py36h516909a_0
pyzmq=18.0.0=py36he6710b0_0
qt=5.9.7=h5867ecd_1
qtconsole=4.4.3=py36_0
readline=8.0=hf8c457e_0
scikit-learn=0.20.1=py36h4989274_0
scipy=1.1.0=py36hfa4b5c9_1
see=1.4.1=pypi_0
send2trash=1.5.0=py36_0
setuptools=41.0.1=py36_0
sip=4.19.8=py36hf484d3e_0
six=1.12.0=py36_1000
snappy=1.1.7=hbae5bb6_3
sqlite=3.29.0=hcee41ef_0
tabulate=0.8.3=py36_0
terminado=0.8.2=py36_0
testpath=0.4.2=py36_0
tk=8.6.9=hed695b0_1002
torchvision-cpu=0.2.2=py_3
tornado=5.1.1=py36h7b6447c_0
traitlets=4.3.2=py36_0
wcwidth=0.1.7=py36_0
webencodings=0.5.1=py36_1
wheel=0.33.4=py36_0
widgetsnbextension=3.4.2=py36_0
xz=5.2.4=h14c3975_1001
yaml=0.1.7=h14c3975_1001
zeromq=4.3.1=he6710b0_3
zlib=1.2.11=h516909a_1005
zstd=1.3.7=h0b5b093_0
